issue #1 attempted solution
